requirements.txt
flask
click>=8.1.7
python-slugify>=8.0.4
pyyaml>=6.0.2
requests>=2.32.3
openai>=1.40.0
python-dotenv>=1.0.1